requirements.txt
Django==5.1.2
djangorestframework==3.15.2
rest_framework_simplejwt==5.3.1
djongo==1.2.31
pymongo==4.11.3
google-generativeai==0.8.4
requests==2.32.3
pandas==2.2.3
python-dotenv==1.0.1
drf-yasg==1.21.7
PyYAML==6.0.1  
